feat: useIndicator 리팩토링
