+++ apps/client/components/Common/index.ts
@@ -9,6 +9,7 @@ export { default as ComponentWithLabel } from "./ComponentWithLabel";
 export { default as Badge } from "./Badge";
 export { default as Gallery } from "./Gallery";
 export { default as HeartWithNumber } from "./HeartWithNumber";
+export { default as PercentBar } from "./PercentBar";
 export * from "./RoundedTag";
 export * from "./Button";
 export * from "./TextField";
